refactor(core): split operations and utils
