--- libs/solana-dapp/account/src/lib/operations/index.ts
@@ -1,2 +1,0 @@
-export * from './get-market-mint-addresses';
-export * from './get-market-orderbook-addresses';
