+++ libs/solana-dapp/account/src/lib/utils/get-orderbook-addresses.ts
@@ -1,6 +1,6 @@
 import { MarketAccount } from '@nx-dapp/solana-dapp/account';
 
-export const getMarketOrderbookAddresses = (
+export const getOrderbookAddresses = (
   marketAccount: MarketAccount
 ): string[] => [
   marketAccount.info.asks.toBase58(),
