+++ libs/solana-dapp/market/src/lib/get-prices.ts
@@ -2,9 +2,9 @@ import { getMarketsData, getMintAccounts } from '@nx-dapp/solana-dapp/account';
 import { Connection } from '@solana/web3.js';
 import { switchMap } from 'rxjs/operators';
 
-import { getMarketAddresses } from './operations';
 import { mapToPrices } from './operators';
 import { GetPricesConfig } from './types';
+import { getMarketAddresses } from './utils';
 
 export const getPrices = (config: GetPricesConfig) => {
   const connection = new Connection(config.rpcEndpoint, 'recent');
