+++ libs/solana-dapp/market/src/lib/operations/index.ts
@@ -1,4 +1,1 @@
 export * from './calculate-best-bid-offer';
-export * from './get-market-by-base-mint';
-export * from './get-market-addresses';
-export * from './get-token-price-by-mint';
