+++ libs/solana-dapp/wallet/utils/get-balances/src/lib/operations/index.ts
@@ -1,3 +1,2 @@
 export * from './calculate-lamports';
 export * from './calculate-quantity';
-export * from './calculate-balance';
