+++ libs/solana-dapp/wallet/utils/get-balances/src/lib/operators/map-to-balance.ts
@@ -3,13 +3,13 @@ import { Balance } from '@nx-dapp/solana-dapp/wallet/types';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
-import { calculateBalance } from '../operations';
+import { createBalance } from '../utils';
 
 export const mapToBalance =
   (userAccounts: TokenAccount[]) =>
   (source: Observable<MintTokenAccount | null>): Observable<Balance | null> =>
     source.pipe(
       map((mintAccount) =>
-        mintAccount ? calculateBalance(userAccounts, mintAccount) : null
+        mintAccount ? createBalance(userAccounts, mintAccount) : null
       )
     );
